feat(pulse): unify item selection with IRT-derived gates and lightweight leapfrog
Replace the 3-band allocation system (frontier/current/review) with a single utility-based item selection function: `utility = information(θ_eff, a, β) × urgency`. Band labels are now emergent from student state, not allocation targets.

Key changes:

- **IRT-derived gates**: Gates are derived directly from θ+σ via `derive_gate_from_irt()` checking P(correct) at reference βs. Removes stability multipliers (×2.5/×1.5/×0.5) — gates can only advance or hold, never regress on a single eval.

- **Continuous response weights**: Replace binary threshold (score >= 9.0) with `response_weight = score / 10.0` in Bayesian EAP likelihood and MLE β estimation. A score of 8.5 now produces a mostly-correct likelihood proportionally.

- **Lightweight leapfrog**: Leapfrog now seeds only competency docs (credibility=0.1) for unlock propagation. No fabricated θ/σ/lifecycle/ability docs — the unified selector handles unlocked skills naturally via high Fisher information.

- **Calibration prior fix**: ProblemTypeRegistry β scale changed from 1.5-8.0 to 1.0-4.0. G4 mastery (P≥0.90 at hardest mode) is now mathematically reachable at θ≈5.7 for gifted students.

- **Deferred write optimization**: Added `defer_session_save`, `defer_competency`, `defer_primitive_history`, `prefetched_session`, `prefetched_global_pass_rate`, and `item_calibration_cache` parameters to minimize Firestore round-trips for batch callers. Session doc loaded once, saved once per session.

- **InMemoryFirestoreService**: New in-memory test harness achieves 500x speedup (31ms vs 10-30s per 15-item session) enabling rapid iteration on selection algorithms and gate thresholds.

- **CalibrationSimulator.tsx**: Updated retention tab to mirror `derive_gate_from_irt()` with per-item theta/sigma/gate progression chart.

+++ backend/app/models/mastery_lifecycle.py
@@ -80,9 +80,9 @@ class MasteryLifecycle(BaseModel):
     # Completion factor (actuarial model — PRD Section 4)
     completion_pct: float = Field(default=0.0, ge=0.0, le=1.0)
 
-    # Practice-mode pass/fail accounting (Gates 2-4 only)
-    passes: int = Field(default=0, ge=0)
-    fails: int = Field(default=0, ge=0)
+    # Practice-mode pass/fail accounting — continuous weights (score/10)
+    passes: float = Field(default=0.0, ge=0.0)
+    fails: float = Field(default=0.0, ge=0.0)
     subskill_pass_rate: float = Field(default=0.0, ge=0.0, le=1.0)
     blended_pass_rate: float = Field(default=0.0, ge=0.0, le=1.0)
     credit_per_pass: float = Field(default=0.25, ge=0.0, le=0.25)
